requirements.txt
charset-normalizer==3.0.1
Django==2.2.28
django-environ==0.9.0
django-registration-redux==2.2
idna==3.4
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
wincertstore==0.2
